+++ ChangeLog.md
@@ -1,5 +1,8 @@
 2020-03
 
+- hmm.learn: extract main function into library.
+  NOTE: I had a bug before: was skipping the first given sequence!
+
 - vq.quantize: extract main function into library
 
 2019-09